+++ lib/src/miui_button.dart
@@ -3,9 +3,13 @@ import 'package:flutter/material.dart';
 class MIUIButton extends StatelessWidget {
   MIUIButton({
     Key key,
+    @required this.child,
     @required this.onPressed,
+    this.colored = true,
   }) : super(key: key);
+  final Widget child;
   final VoidCallback onPressed;
+  final bool colored;
   @override
   Widget build(BuildContext context) {
     return MaterialButton(
@@ -14,12 +18,16 @@ class MIUIButton extends StatelessWidget {
       disabledElevation: 0,
       highlightElevation: 0,
       hoverElevation: 0,
-      color: Colors.blueAccent,
-      splashColor: Color(0xff0099FF),
-      textColor: Colors.white,
+      color: colored ? Theme.of(context).primaryColor : null,
+      splashColor: colored ? Color(0xff0099FF) : null,
+      textColor: colored
+          ? Color(0xddffffff -
+              Theme.of(context).textTheme.title.color.value +
+              0xdd000000)
+          : null,
       padding: EdgeInsets.only(top: 14, bottom: 14),
       onPressed: this.onPressed,
-      child: Text('test'),
+      child: this.child,
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(40),
       ),