+++ lib/src/miui_menu.dart
@@ -1,15 +1,6 @@
 import 'package:flutter/material.dart';
 
-showMIUIMenu({
-  @required BuildContext context,
-  List<PopupMenuItem> items,
-  RelativeRect rect,
-}) {
-  showMenu(
-      context: context,
-      position: rect,
-      items: items,
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(15),
-      ));
-}
+// ignore: non_constant_identifier_names
+ShapeBorder MIUIMenuShape = RoundedRectangleBorder(
+  borderRadius: BorderRadius.circular(15),
+);