+++ lib/src/miui_scroll_physics.dart
@@ -6,14 +6,3 @@ class MIUIScrollPhysics extends ScrollPhysics {
     return AlwaysScrollableScrollPhysics(parent: BouncingScrollPhysics());
   }
 }
-
-class MIUIDivider extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Divider(
-      height: 1,
-      indent: 20,
-      endIndent: 20,
-    );
-  }
-}