+++ CHANGELOG.md
@@ -1,7 +1,22 @@
-## [0.0.1] - 2019.12.12.
+## [0.0.5] - 2020.02.05
 
-* initial release.
+* [menu] use shape to change Material Menu Shape.
+* [dialog] fix color.
+* [listTile] use custom Widget to change Material ListTile's subtitle and trialing.
+* delete tag and textField.
+
+## [0.0.4] - 2019.12.25
+
+* small fix.
+
+## [0.0.3] - 2019.12.25
+
+* add button,use multi files.
 
 ## [0.0.2] - 2019.12.14
 
 * add some widget.
+
+## [0.0.1] - 2019.12.12.
+
+* initial release.